pyproject.toml
[tool.poetry]
name = "social-systems-delivery"
version = "0.1.0"
description = ""
authors = ["krasniy-1889 <neytirri@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = "^0.32.0"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
loguru = "^0.7.2"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
psycopg-binary = "^3.2.3"
greenlet = "^3.1.1"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
psycopg = "^3.2.3"
httpx = "^0.28.0"
cuid = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
